Move wear packages to new package
Having it the same was causing conflicts with the mobile app

Also delete old colors object
