+++ wear/src/main/AndroidManifest.xml
@@ -9,7 +9,7 @@
 
     <application
         android:allowBackup="true"
-        android:name=".test.PowerHourWearApp"
+        android:name=".wear.PowerHourWearApp"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:supportsRtl="true"
@@ -19,7 +19,7 @@
             android:required="false" />
 
         <activity
-            android:name=".Game.GameActivity"
+            android:name=".wear.Game.GameActivity"
             android:label="@string/app_name"
             android:launchMode="singleTask"
             android:theme="@android:style/Theme.DeviceDefault.Light">
@@ -31,7 +31,7 @@
         </activity>
 
         <service
-            android:name=".DataLayer.GameListenerService"
+            android:name=".wear.DataLayer.GameListenerService"
             android:enabled="true"
             android:exported="true">
             <intent-filter>
@@ -40,7 +40,7 @@
         </service>
 
         <activity
-            android:name=".Game.FinishActivity"
+            android:name=".wear.Game.FinishActivity"
             android:label="@string/title_activity_finish"
             android:launchMode="singleTask"
             android:theme="@android:style/Theme.DeviceDefault.Light" />
