+++ wear/src/main/java/ca/hoogit/powerhour/wear/DataLayer/GoogleApiManager.java
@@ -15,7 +15,7 @@
  * with this program; if not, write to the Free Software Foundation, Inc.,
  * 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
  */
-package ca.hoogit.powerhour.DataLayer;
+package ca.hoogit.powerhour.wear.DataLayer;
 
 import android.content.Context;
 import android.os.Bundle;
