+++ wear/src/main/java/ca/hoogit/powerhour/wear/Game/GameActivity.java
@@ -1,6 +1,5 @@
-package ca.hoogit.powerhour.Game;
+package ca.hoogit.powerhour.wear.Game;
 
-import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
@@ -9,7 +8,6 @@ import android.support.wearable.activity.WearableActivity;
 import android.support.wearable.view.WatchViewStub;
 import android.util.Log;
 import android.widget.RelativeLayout;
-import android.widget.Toast;
 
 import com.google.android.gms.common.ConnectionResult;
 import com.google.android.gms.common.api.GoogleApiClient;
@@ -21,10 +19,9 @@ import com.google.android.gms.wearable.MessageApi;
 import com.google.android.gms.wearable.MessageEvent;
 import com.google.android.gms.wearable.Wearable;
 
-import ca.hoogit.powerhour.DataLayer.GameInformation;
-import ca.hoogit.powerhour.DataLayer.Message;
+import ca.hoogit.powerhour.wear.DataLayer.GameInformation;
+import ca.hoogit.powerhour.wear.DataLayer.Message;
 import ca.hoogit.powerhour.R;
-import ca.hoogit.powerhour.Utils.Colors;
 import ca.hoogit.powerhourshared.DataLayer.Consts;
 
 public class GameActivity extends WearableActivity implements
