+++ wear/src/main/res/layout/activity_finish.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:gravity="center"
-    tools:context=".Game.FinishActivity"
+    tools:context=".wear.Game.FinishActivity"
     tools:deviceIds="wear">
 
     <RelativeLayout
